@@ -19,7 +19,8 @@
     "lint": "eslint --ignore-path=.eslintignore --ext .js,.jsx .",
     "lint-fix": "eslint --fix --ignore-path=.eslintignore --ext .js,.jsx .",
     "sync-backend": "babel-node --presets env src/syncBackend.js",
-    "cypress": "cypress"
+    "cypress": "cypress",
+    "cypress-debug": "cypress open --config watchForFileChanges=true"
   },
   "repository": {
     "type": "git",
@@ -80,7 +81,6 @@
     "jquery": "3.1.1",
     "json-bigint": "^0.3.0",
     "lodash": "^4.17.11",
-    "lodash.throttle": "^4.1.1",
     "mapbox-gl": "^0.45.0",
     "mathjs": "^3.20.2",
     "moment": "^2.20.1",
@@ -138,13 +138,14 @@
     "babel-loader": "^7.1.4",
     "babel-plugin-css-modules-transform": "^1.1.0",
     "babel-plugin-dynamic-import-node": "^1.2.0",
+    "babel-plugin-lodash": "^3.3.4",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
     "babel-polyfill": "^6.23.0",
     "babel-preset-airbnb": "^2.1.1",
     "babel-preset-env": "^1.7.0",
     "chai": "^4.0.2",
     "clean-webpack-plugin": "^0.1.19",
-    "css-loader": "^0.28.0",
+    "css-loader": "^1.0.0",
     "cypress": "^3.0.3",
     "enzyme": "^3.3.0",
     "enzyme-adapter-react-16": "^1.1.1",
@@ -164,7 +165,6 @@
     "imports-loader": "^0.7.1",
     "istanbul": "^1.0.0-alpha",
     "jsdom": "9.12.0",
-    "json-loader": "^0.5.4",
     "less": "^2.6.1",
     "less-loader": "^4.1.0",
     "mini-css-extract-plugin": "^0.4.0",
@@ -178,13 +178,15 @@
     "react-test-renderer": "^15.6.2",
     "redux-mock-store": "^1.2.3",
     "sinon": "^4.5.0",
+    "speed-measure-webpack-plugin": "^1.2.3",
     "style-loader": "^0.21.0",
+    "terser-webpack-plugin": "^1.1.0",
     "transform-loader": "^0.2.3",
-    "uglifyjs-webpack-plugin": "^1.1.0",
     "url-loader": "^1.0.1",
-    "webpack": "^4.18.0",
+    "webpack": "^4.19.0",
     "webpack-assets-manifest": "^3.0.1",
-    "webpack-cli": "^2.0.10",
+    "webpack-bundle-analyzer": "^3.0.2",
+    "webpack-cli": "^3.1.1",
     "webpack-dev-server": "^3.1.7",
     "webpack-sources": "^1.1.0"
   }
