@@ -8,16 +8,31 @@ import sinon from 'sinon';
 
 import App from '../../../src/SqlLab/components/App';
 import TabbedSqlEditors from '../../../src/SqlLab/components/TabbedSqlEditors';
-import { sqlLabReducer } from '../../../src/SqlLab/reducers';
+import getInitialState from '../../../src/SqlLab/getInitialState';
 
-describe('App', () => {
+describe('SqlLab App', () => {
   const middlewares = [thunk];
   const mockStore = configureStore(middlewares);
-  const store = mockStore({ sqlLab: sqlLabReducer(undefined, {}), messageToasts: [] });
-
+  let store;
   let wrapper;
+
+  before(() => {
+    const bootstrapData = {
+      common: {
+        feature_flags: {
+          FOO_BAR: true,
+        },
+      },
+    };
+    store = mockStore(getInitialState(bootstrapData), {});
+  });
+
   beforeEach(() => {
-    wrapper = shallow(<App />, { context: { store } }).dive();
+    wrapper = shallow(<App />, { context: { store } });
+  });
+
+  it('should set feature flags', () => {
+    expect(wrapper.prop('isFeatureEnabled')('FOO_BAR')).to.equal(true);
   });
 
   it('is valid', () => {
@@ -25,14 +40,16 @@ describe('App', () => {
   });
 
   it('should handler resize', () => {
-    sinon.spy(wrapper.instance(), 'getHeight');
-    wrapper.instance().handleResize();
-    expect(wrapper.instance().getHeight.callCount).to.equal(1);
-    wrapper.instance().getHeight.restore();
+    const inner = wrapper.dive();
+    sinon.spy(inner.instance(), 'getHeight');
+    inner.instance().handleResize();
+    expect(inner.instance().getHeight.callCount).to.equal(1);
+    inner.instance().getHeight.restore();
   });
 
   it('should render', () => {
-    expect(wrapper.find('.SqlLab')).to.have.length(1);
-    expect(wrapper.find(TabbedSqlEditors)).to.have.length(1);
+    const inner = wrapper.dive();
+    expect(inner.find('.SqlLab')).to.have.length(1);
+    expect(inner.find(TabbedSqlEditors)).to.have.length(1);
   });
 });
