@@ -2,9 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { OverlayTrigger, Popover } from 'react-bootstrap';
 import { SketchPicker } from 'react-color';
-
 import ControlHeader from '../ControlHeader';
-import { bnbColors } from '../../../modules/colors';
+import { getScheme } from '../../../modules/ColorSchemeManager';
 
 const propTypes = {
   onChange: PropTypes.func,
@@ -59,7 +58,7 @@ export default class ColorPickerControl extends React.Component {
         <SketchPicker
           color={this.props.value}
           onChange={this.onChange}
-          presetColors={bnbColors.filter((s, i) => i < 7)}
+          presetColors={getScheme().filter((s, i) => i < 7)}
         />
       </Popover>);
   }
