@@ -1,9 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-
+import shortid from 'shortid';
 import { XYChart, AreaSeries, CrossHair, LinearGradient } from '@data-ui/xy-chart';
-
-import { brandColor } from '../../modules/colors';
+import { BRAND_COLOR } from '../../modules/colors';
 import { formatDateVerbose } from '../../modules/dates';
 import { computeMaxFontSize } from '../../modules/visUtils';
 
@@ -51,30 +50,33 @@ const propTypes = {
   bigNumber: PropTypes.number.isRequired,
   formatBigNumber: PropTypes.func,
   subheader: PropTypes.string,
-  showTrendline: PropTypes.bool,
+  showTrendLine: PropTypes.bool,
   startYAxisAtZero: PropTypes.bool,
-  trendlineData: PropTypes.array,
+  trendLineData: PropTypes.array,
   mainColor: PropTypes.string,
-  gradientId: PropTypes.string,
   renderTooltip: PropTypes.func,
 };
 const defaultProps = {
   className: '',
   formatBigNumber: identity,
   subheader: '',
-  showTrendline: false,
+  showTrendLine: false,
   startYAxisAtZero: true,
-  trendlineData: null,
-  mainColor: brandColor,
-  gradientId: '',
+  trendLineData: null,
+  mainColor: BRAND_COLOR,
   renderTooltip: renderTooltipFactory(identity),
 };
 
 class BigNumberVis extends React.Component {
+  constructor(props) {
+    super(props);
+    this.gradientId = shortid.generate();
+  }
+
   getClassName() {
-    const { className, showTrendline } = this.props;
+    const { className, showTrendLine } = this.props;
     const names = `big_number ${className}`;
-    if (showTrendline) {
+    if (showTrendLine) {
       return names;
     }
     return `${names} no_trendline`;
@@ -159,11 +161,10 @@ class BigNumberVis extends React.Component {
   renderTrendline(maxHeight) {
     const {
       width,
-      trendlineData,
+      trendLineData,
       mainColor,
       subheader,
       renderTooltip,
-      gradientId,
       startYAxisAtZero,
     } = this.props;
     return (
@@ -181,13 +182,13 @@ class BigNumberVis extends React.Component {
         snapTooltipToDataX
       >
         <LinearGradient
-          id={gradientId}
+          id={this.gradientId}
           from={mainColor}
           to="#fff"
         />
         <AreaSeries
-          data={trendlineData}
-          fill={`url(#${gradientId})`}
+          data={trendLineData}
+          fill={`url(#${this.gradientId})`}
           stroke={mainColor}
         />
         <CrossHair
@@ -203,10 +204,10 @@ class BigNumberVis extends React.Component {
   }
 
   render() {
-    const { showTrendline, height } = this.props;
+    const { showTrendLine, height } = this.props;
     const className = this.getClassName();
 
-    if (showTrendline) {
+    if (showTrendLine) {
       const chartHeight = Math.floor(PROPORTION.TRENDLINE * height);
       const allTextHeight = height - chartHeight;
       return (
