@@ -160,7 +160,7 @@ class TableModelView(DatasourceModelView, DeleteMixin, YamlExportMixin):  # noqa
     edit_title = _('Edit Table')
 
     list_columns = [
-        'link', 'database',
+        'link', 'database_name',
         'changed_by_', 'modified']
     order_columns = ['modified']
     add_columns = ['database', 'schema', 'table_name']
@@ -229,6 +229,7 @@ class TableModelView(DatasourceModelView, DeleteMixin, YamlExportMixin):  # noqa
         'link': _('Table'),
         'changed_by_': _('Changed By'),
         'database': _('Database'),
+        'database_name': _('Database'),
         'changed_on_': _('Last Changed'),
         'filter_select_enabled': _('Enable Filter Select'),
         'schema': _('Schema'),
