@@ -6,11 +6,6 @@ Revises: 956a063c52b3
 Create Date: 2016-05-27 15:03:32.980343
 
 """
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-from __future__ import unicode_literals
-
 # revision identifiers, used by Alembic.
 revision = '1226819ee0e3'
 down_revision = '956a063c52b3'
@@ -33,10 +28,11 @@ def find_constraint_name(upgrade=True):
 
 def upgrade():
     try:
-        constraint = find_constraint_name() or 'fk_columns_column_name_datasources'
+        constraint = find_constraint_name()
         with op.batch_alter_table("columns",
                 naming_convention=naming_convention) as batch_op:
-            batch_op.drop_constraint(constraint, type_="foreignkey")
+            if constraint:
+                batch_op.drop_constraint(constraint, type_="foreignkey")
             batch_op.create_foreign_key(
                 'fk_columns_datasource_name_datasources',
                 'datasources',
